# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Restaurantes",
        "description": "Listado de endpoints de la api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/table": {
            "post": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Crear una tabla nueva",
                "description": "Agrega una nueva mesa.",
                "operationId": "6ad44abbe32bb50b49e661ecd57f732f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nro": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Mesa guardada"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/change/{id}": {
            "put": {
                "tags": [
                    "Mesas"
                ],
                "summary": "Cambiar disponibilidad de una mesa por su id",
                "description": "Cambia el estado de la mesa de false a true o viceversa.",
                "operationId": "0fea71565e4b2ee1b52f363d4ab167ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Estado cambiado correctamente"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}